requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

gdown
gradio
numpy
opencv-python
pandas
patchify
pillow
torch
torchvision
tqdm
requests
joblib